# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Management API",
        "description": "Современное API для управления автомобилями с возможностью совместного использования. Построено на Laravel с использованием Clean Architecture, Repository Pattern и Service Layer.",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Локальный сервер разработки"
        },
        {
            "url": "http://api.test29.test/api",
            "description": "Сервер разработки (Herd)"
        }
    ],
    "paths": {
...
        "/cars/{id}/attach": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Прикрепить автомобиль к текущему пользователю",
                "description": "Добавляет связь между автомобилем и текущим пользователем",
                "operationId": "333abd43f75f474a99a3b6923928e459",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно прикреплен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car attached successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "404": {
                        "description": "Автомобиль не найден"
                    },
                    "409": {
                        "description": "Автомобиль уже прикреплен"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars/{id}/detach": {
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Открепить автомобиль от текущего пользователя",
                "description": "Удаляет связь между автомобилем и текущим пользователем",
                "operationId": "6414b1cfa5b7a70ab648f06000a6fb85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно откреплен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car detached successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "404": {
                        "description": "Автомобиль не найден или не прикреплен"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars/{id}/users": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Получить список пользователей автомобиля",
                "description": "Возвращает всех пользователей, которые имеют доступ к автомобилю",
                "operationId": "23d4b32a179ee2d66ad1bca7f7b4776a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список пользователей автомобиля",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "car": {
                                                    "$ref": "#/components/schemas/Car"
                                                },
                                                "users": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Автомобиль не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cars/{id}/share": {
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Поделиться автомобилем с другим пользователем",
                "description": "Добавляет другого пользователя к автомобилю для совместного использования",
                "operationId": "cfc2681441dae917e58f23aec46e48ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID автомобиля",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_email"
                                ],
                                "properties": {
                                    "user_email": {
                                        "description": "Email пользователя для добавления",
                                        "type": "string",
                                        "format": "email",
                                        "example": "friend@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Автомобиль успешно расшарен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Car shared successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    },
                    "403": {
                        "description": "Доступ запрещен"
                    },
                    "404": {
                        "description": "Автомобиль или пользователь не найден"
                    },
                    "409": {
                        "description": "Пользователь уже имеет доступ к автомобилю"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Car": {
                "title": "Автомобиль",
                "description": "Модель автомобиля",
                "properties": {
                    "id": {
                        "description": "ID автомобиля",
                        "type": "integer",
                        "example": 1
                    },
                    "car_brand_id": {
                        "description": "ID бренда",
                        "type": "integer",
                        "example": 1
                    },
                    "car_model_id": {
                        "description": "ID модели",
                        "type": "integer",
                        "example": 1
                    },
                    "year": {
                        "description": "Год выпуска",
                        "type": "string",
                        "example": "2020",
                        "nullable": true
                    },
                    "color": {
                        "description": "Цвет",
                        "type": "string",
                        "example": "Красный",
                        "nullable": true
                    },
                    "mileage": {
                        "description": "Пробег",
                        "type": "integer",
                        "example": 50000,
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Дата обновления",
                        "type": "string",
                        "format": "date-time"
                    },
                    "car_brand": {
                        "$ref": "#/components/schemas/CarBrand"
                    },
                    "car_model": {
                        "$ref": "#/components/schemas/CarModel"
                    },
                    "users": {
                        "description": "Пользователи, связанные с автомобилем",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
            "CarBrand": {
                "title": "Бренд автомобиля",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "CarModel": {
                "title": "Модель автомобиля",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "car_brand_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Camry"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "car_brand": {
                        "$ref": "#/components/schemas/CarBrand"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "Пользователь",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Введите токен в формате: Bearer {your-token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}